Fixed Training & Range
Guns use Combat Mastery as the TargetAction skill, but the tAction
object in aura sets the AttackerAction skill as what the target would be
getting hit by. In this case, checking for combat mastery would give a
green flag to ALL gunner skills, and checking for the AttackerSkill
wouldn't work. Solution: Explicitly set AttackerSkill.